changed price variable to decimal

+++ Applikation med API/Models/Product.cs
@@ -8,8 +8,8 @@ namespace Applikation_med_API.Models
         [Key]
         public int ID { get; set; }
         public string Name { get; set; }
-        public string? ImagePath { get; set; } 
-        public double Price { get; set; }
+        public string? ImagePath { get; set; }
+        public decimal Price { get; set; }
         public string? Category { get; set; }
         public string? Description { get; set; }
     }
